Reorganize workspace quota and history pages
