+++ web/src/index.css
@@ -3900,6 +3900,9 @@ body,
 
 .router-service-pricing-history-link {
     color: #2185d0;
+    font-size: 14px;
+    line-height: 1.4;
+    font-weight: 400;
     text-decoration: none;
 }
