+++ web/src/locales/zh/translation.json
@@ -153,7 +153,7 @@
       }
     },
     "payment_history": {
-      "button": "支付记录",
+      "button": "支付历史",
       "title": "支付历史",
       "status_title": "支付状态",
       "back_to_pricing": "返回订阅"
